split registerPush to its own file
