+++ src/workers/client.js
@@ -1,5 +1,4 @@
 import urlBase64ToUint8Array from './urlBase64ToUint8Array';
-import { postSubscription } from '../services/subscribe';
 const publicVapidKey = 'BBDk-M7Qx45gAgYzwp3epWc_PK7Iv4cF9kehNJ0YWbdLinh8JFVUhPgvoq2fQfQYLkqKISqc55D1z5Shnqk6M8A';
 
 if('serviceWorker' in navigator) {
@@ -16,15 +15,4 @@ async function run() {
   const registration = await navigator.serviceWorker.register('/src/workers/worker.js', { scope: '/src/workers/' })
     .then(registration);
   console.log('Registered service worker');
-
-  console.log('Registering push');
-  const subscription = await registration.pushManager.subscribe({
-    userVisibleOnly: true,
-    applicationServerKey: urlBase64ToUint8Array(publicVapidKey)
-  });
-  console.log('Registered push');
-
-  console.log('Sending push');
-  postSubscription(subscription);
-  console.log('Sent push');
 }
